// server/src/relay/client/relay-client.ts
export interface RelayTorrent {
  /** checking_files=1, downloading_metadata=2, downloading=3, finished=4, seeding=5, unused_enum_for_backwards_compatibility_allocating=6, checking_resume_data=7 https://www.libtorrent.org/reference-Torrent_Status.html#torrent_status */
  state: number;
  progress: number;
  name: string;
  infoHash: string;
  downloadSpeed: number;
  uploadSpeed: number;
  downloaded: number;
  uploaded: number;
  total: number;
  maxConnections: number;
  connections: number;
}

export interface RelayTorrentState {
  /** checking_files=1, downloading_metadata=2, downloading=3, finished=4, seeding=5, unused_enum_for_backwards_compatibility_allocating=6, checking_resume_data=7 https://www.libtorrent.org/reference-Torrent_Status.html#torrent_status */
  state: number;
  progress: number;
}

export interface UpdateRelayTorrent {
  downloadFullTorrent?: boolean | null;
}

export interface UpdateSettings {
  downloadLimit?: number | null;
  uploadLimit?: number | null;
  port?: number | null;
  connectionsLimit?: number | null;
  torrentConnectionsLimit?: number | null;
  enableUpnpAndNatpmp?: boolean | null;
}