// server/src/relay/client/models/UpdateSettings.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateSettings = {
    downloadLimit?: (number | null);
    uploadLimit?: (number | null);
    port?: (number | null);
    connectionsLimit?: (number | null);
    torrentConnectionsLimit?: (number | null);
    enableUpnpAndNatpmp?: (boolean | null);
};